экран выбора помещения

+++ app/src/main/java/com/example/healthyapp/db/model/entity/Placement.kt
@@ -2,8 +2,8 @@ package com.example.healthyapp.db.model.entity
 
 data class Placement(
     val id: String,
-    val number: Int,
-    val length: Int,
-    val height: Int,
-    val width: Int
+    val number: Long,
+    val length: Long,
+    val height: Long,
+    val width: Long
 )
